LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/percentage-height-in-flexbox-expected.txt



PASS #target offsetSize matches #container offsetSize
FAIL #target offsetSize matches #container offsetSize after resize assert_equals: expected 50 but got 100
